+++ src/Either.ts
@@ -241,7 +241,7 @@ export const stringifyJSON = (u: unknown): Either<unknown, string> => tryCatch((
  * @category constructors
  * @since 3.0.0
  */
-export const fromOption: <E>(onNone: Lazy<E>) => <A>(ma: Option<A>) => Either<E, A> = (onNone) => (ma) =>
+export const fromOption = <E>(onNone: Lazy<E>) => <A>(ma: Option<A>): Either<E, A> =>
   ma._tag === 'None' ? left(onNone()) : right(ma.value)
 
 /**
@@ -253,8 +253,7 @@ export const fromOption: <E>(onNone: Lazy<E>) => <A>(ma: Option<A>) => Either<E,
  *   pipe(
  *     1,
  *     fromPredicate(
- *       (n) => n > 0,
- *       () => 'error'
+ *       (n) => n > 0
  *     )
  *   ),
  *   right(1)
@@ -263,20 +262,19 @@ export const fromOption: <E>(onNone: Lazy<E>) => <A>(ma: Option<A>) => Either<E,
  *   pipe(
  *     -1,
  *     fromPredicate(
- *       (n) => n > 0,
- *       () => 'error'
+ *       (n) => n > 0
  *     )
  *   ),
- *   left('error')
+ *   left(-1)
  * )
  *
  * @category constructors
  * @since 3.0.0
  */
 export const fromPredicate: {
-  <A, B extends A, E>(refinement: Refinement<A, B>, onFalse: (a: A) => E): (a: A) => Either<E, B>
-  <A, E>(predicate: Predicate<A>, onFalse: (a: A) => E): (a: A) => Either<E, A>
-} = <A, E>(predicate: Predicate<A>, onFalse: (a: A) => E) => (a: A) => (predicate(a) ? right(a) : left(onFalse(a)))
+  <A, B extends A>(refinement: Refinement<A, B>): (a: A) => Either<A, B>
+  <A>(predicate: Predicate<A>): (a: A) => Either<A, A>
+} = <A>(predicate: Predicate<A>) => (a: A) => (predicate(a) ? right(a) : left(a))
 
 // -------------------------------------------------------------------------------------
 // destructors
@@ -1293,7 +1291,7 @@ export const traverseArrayWithIndex = <E, A, B>(f: (index: number, a: A) => Eith
  * @example
  *
  *
- * import { traverseArray, left, right, fromPredicate } from 'fp-ts/Either'
+ * import * as E from 'fp-ts/Either'
  * import { pipe } from 'fp-ts/function'
  * import * as A from 'fp-ts/ReadonlyArray'
  *
@@ -1301,21 +1299,20 @@ export const traverseArrayWithIndex = <E, A, B>(f: (index: number, a: A) => Eith
  * assert.deepStrictEqual(
  *   pipe(
  *     arr,
- *     traverseArray((x) => right(x))
+ *     E.traverseArray(E.right)
  *   ),
- *   right(arr)
+ *   E.right(arr)
  * )
  * assert.deepStrictEqual(
  *   pipe(
  *     arr,
- *     traverseArray(
- *       fromPredicate(
- *         (x) => x > 5,
- *         () => 'a'
+ *     E.traverseArray(
+ *       E.fromPredicate(
+ *         (x) => x > 5
  *       )
  *     )
  *   ),
- *   left('a')
+ *   E.left(0)
  * )
  * @since 3.0.0
  */
@@ -1329,13 +1326,13 @@ export const traverseArray: <E, A, B>(
  *
  * @example
  *
- * import { sequenceArray, left, right } from 'fp-ts/Either'
+ * import * as E from 'fp-ts/Either'
  * import { pipe } from 'fp-ts/function'
  * import * as A from 'fp-ts/ReadonlyArray'
  *
  * const arr = A.range(0, 10)
- * assert.deepStrictEqual(pipe(arr, A.map(right), sequenceArray), right(arr))
- * assert.deepStrictEqual(pipe(arr, A.map(right), A.cons(left('Error')), sequenceArray), left('Error'))
+ * assert.deepStrictEqual(pipe(arr, A.map(E.right), E.sequenceArray), E.right(arr))
+ * assert.deepStrictEqual(pipe(arr, A.map(E.right), A.cons(E.left('Error')), E.sequenceArray), E.left('Error'))
  *
  * @since 3.0.0
  */